Expand getPluginData
