+++ tools/cli-core/package.json
@@ -15,10 +15,11 @@
 		"typescript": "^4.8.3"
 	},
 	"dependencies": {
+		"camel-case": "^4.1.2",
 		"chalk": "^4.1.2",
 		"dotenv": "^10.0.0",
 		"ora": "^5.4.1",
-		"simple-git": "^3.10.0",
+		"simple-git": "^3.14.0",
 		"ts-node": "^10.9.1",
 		"uuid": "^8.3.2"
 	}
