+++ tools/version-bump/package.json
@@ -17,8 +17,8 @@
 	"dependencies": {
 		"@commander-js/extra-typings": "^0.1.0",
 		"chalk": "^4.1.2",
-		"commander": "9.4.0",
 		"cli-core": "workspace:*",
+		"commander": "9.4.0",
 		"express": "^4.18.1",
 		"ora": "^5.4.1",
 		"semver": "^7.3.2",
